Only move existing overlays

+++ src/usr/lib/rsetup/cli/u-boot-menu.sh
@@ -46,7 +46,10 @@ disable_overlays() {
         mapfile -t RSETUP_MANAGED_OVERLAYS < "$U_BOOT_FDT_OVERLAYS_DIR/managed.list"
         for i in "${RSETUP_MANAGED_OVERLAYS[@]}"
         do
-            mv -- "$U_BOOT_FDT_OVERLAYS_DIR/$i" "$U_BOOT_FDT_OVERLAYS_DIR/${i}.disabled"
+            if [[ -f "$U_BOOT_FDT_OVERLAYS_DIR/$i" ]]
+            then
+                mv -- "$U_BOOT_FDT_OVERLAYS_DIR/$i" "$U_BOOT_FDT_OVERLAYS_DIR/${i}.disabled"
+            fi
         done
     else
         for i in "$U_BOOT_FDT_OVERLAYS_DIR"/*.dtbo
